pyproject.toml
[project]
name = "convert-config-hook"
version = "0.1.0"
license = { file = "LICENSE" }
description = "Pre-commit hook to convert JSON/TOML/YAML files."
readme = "README.md"
authors = [{ name = "Ralf Gabriels", email = "ralf@gabriels.dev" }]
dependencies = ["pyyaml", "tomli-w", "tomli>=2; python_version < '3.11'"]
requires-python = ">=3.11"

[project.scripts]
convert = "convert:main"
